(BOLT-1380) Use tty-prompt instead of demo-magic to run demo
The demo-magic demo taught us that there's demand throughout Puppet
for
a standalone Bolt demo that can easily be shared and run by non-Bolt
engineers. This is the first step in creating a more robust, usable,
and
enjoyable demo. It includes several changes:

* **Remove the network device demo**. Setting up a network device
  within a local VM infrastructure is difficult, and only a small
  percentage of demo participants care about network device support.
  It isn't worth the difficulty of setup and maintenance to continue
  to support the network device demo.
* **Use tty-prompt instead of demo-magic**. `demo-magic` lacked a
  few support for features we want the demo to have, specifically
  history support (up-arrow should show the last command) and
  validating user input and allowing for retries if incorrect.
  `tty-prompt`, while not explicitly made for demos, does support
  all three of these features.

These both led to other big changes, like installing ruby such that
the
vagrant user can use it in the Bolt controller VM and removing the
all
the network-device setup and demo content.