(BOLT-1447) Use plugin model
This changes the demo writing process to a plugin model. The demo class
is added to a class variable of the demo engine when the demo file is
loaded. Demos and their titles are displayed for the user to select
from, then once a demo is selected the class is instantiated and `run`
is called. This decreases the boilerplate for demos, and demos are only
loaded if 'active' and instantiated if they are used.